# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Series",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        },
        {
            "url": "http://animeStore.test/api/v1"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesion",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@email.com",
                                    "password": "thiswillbeencrypted"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Crea un nuevo usuario",
                "operationId": "App\\Http\\Controllers\\AuthController::signup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "John",
                                    "email": "john@email.com",
                                    "password": "thiswillbeencrypted"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtener el usuario de la sesion actual",
                "operationId": "App\\Http\\Controllers\\AuthController::me",
                "responses": {
                    "201": {
                        "description": "Usuario actual"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/series/{serie}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Obtener los comentarios pertenecientes a una serie ",
                "description": "Regresa la informacion de todos los comentarios pertenecientes a una serie en particular",
                "operationId": "App\\Http\\Controllers\\CommentController::index",
                "parameters": [
                    {
                        "name": "serie",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "404": {
                        "description": "Una o más series pertenecientes al paquete\n     *          no se encuentran o el paquete mismo no se encuentra"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Añadir un comentario",
                "description": "Añade la informacion perteneciente a un comentario en una serie en particular",
                "operationId": "App\\Http\\Controllers\\CommentController::store",
                "parameters": [
                    {
                        "name": "serie",
                        "in": "path",
                        "description": "Serie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "La serie no se encuentra"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Auth": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer Auth": {
                "type": "http",
                "description": "Add bearer token to endpoints",
                "name": "Bearer Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}